clean_version
