+++ ultralytics/tensorleap_folder/.gitattributes
@@ -1,0 +1,2 @@
+*.h5 filter=lfs diff=lfs merge=lfs -text
+*.pt filter=lfs diff=lfs merge=lfs -text
