--- ultralytics/tensorleap_folder/yolov11/leap.yaml
@@ -1,10 +1,0 @@
-codeIntegrationId: 67ce9f07fd3378c34d0eab38
-projectId: 67ceff0d7c39706543d16ead
-secretId: ""
-branch: master
-entryFile: leap_binder.py
-include:
-    - leap_binder.py
-    - leap_custom_test.py
-    - config.py
-    - requirements.txt
