--- ultralytics/tensorleap_folder/yolov11/leap_mapping.yaml
@@ -1,47 +1,0 @@
-decorators:
-  - operation: Input
-    data:
-      type: Input
-      output_name: image
-    id: '144'
-    inputs: {}
-    outputs:
-      image:
-        - inputKey: input
-          operation: ZeroPadding2D
-          name: model.0/conv/Conv_tl_conv_pad_1
-          id: '143'
-  - operation: CustomLoss
-    data:
-      type: CustomLoss
-      name: dummy_loss
-      selected: dummy_loss
-      user_unique_name: dummy_loss
-      arg_names:
-        - pred
-        - gt
-    id: '584'
-    inputs:
-      pred:
-        - outputKey: merged
-          operation: Concatenate
-          name: model.23/Concat_5_tl_concat_2
-          id: '1'
-      gt:
-        - outputKey: classes
-          operation: GroundTruth
-          id: '585'
-    outputs: {}
-  - operation: GroundTruth
-    data:
-      type: GroundTruth
-      output_name: classes
-      selected: classes
-    id: '585'
-    inputs: {}
-    outputs:
-      classes:
-        - inputKey: gt
-          operation: CustomLoss
-          id: '584'
-layers: []
