--- ultralytics/tensorleap_folder/yolov11/requirements.txt
@@ -1,43 +1,0 @@
-#numpy>=1.23.0,<2.1.1
-matplotlib>=3.3.0
-#opencv-python>=4.6.0
-pillow>=7.1.2
-pyyaml>=5.3.1
-requests>=2.23.0
-scipy>=1.4.1
-torch>=1.8.0
-torchvision>=0.9.0
-tqdm>=4.64.0
-#psutil
-py-cpuinfo
-pandas>=1.1.4
-seaborn>=0.11.0
-ultralytics-thop>=2.0.0
-code-loader>=1.0.75,<2.0.0
-tensorflow-macos==2.15.0
-ipython
-pytest
-pytest-cov
-coverage
-mkdocs>=1.6.0
-mkdocs-material>=9.5.9
-mkdocstrings
-mkdocs-redirects
-mkdocs-ultralytics-plugin>=0.1.17
-mkdocs-macros-plugin>=1.0.5
-onnx>=1.12.0
-coremltools>=7.0
-scikit-learn>=1.3.2
-openvino>=2024.0.0,<2025.0.0
-tensorflowjs>=3.9.0
-keras
-flatbuffers>=23.5.26,<100
-h5py!=3.11.0
-#comet
-tensorboard>=2.13.0
-dvclive>=2.12.0
-shapely>=2.0.0
-streamlit
-hub-sdk>=0.0.12
-albumentations>=1.4.6
-pycocotools>=2.0.7
